move DimensionClassParser to o.e.sh.core.types package
